backend/python/app/zabbix_to_word/requirements.txt
appdirs==1.4.3
lxml==3.7.3
packaging==16.8
pyparsing==2.2.0
python-docx==0.8.6
PyYAML==3.12
pyzabbix==0.7.4
requests==2.13.0
six==1.10.0
